chore(data-loader): refactoring and add tests
