+++ packages/data-loader/src/parsers/csvParser/__tests__/formatToKintoneRecords.test.ts
@@ -1,0 +1,31 @@
+import { KintoneFormFieldProperty } from "@kintone/rest-api-client";
+import { formatToKintoneRecords } from "../formatToKintoneRecords";
+
+const records = [
+  {
+    singleLineText: "single line text",
+    multiSelect: "sample1\nsample2",
+  },
+];
+const subTableFieldsJson: {
+  properties: { [k: string]: KintoneFormFieldProperty.OneOf };
+} = require("./fixtures/subtable_fields.json");
+
+describe("formatToKintoneRecords", () => {
+  it("should", () => {
+    expect(
+      formatToKintoneRecords({ records, fieldsJson: subTableFieldsJson })
+    ).toEqual([
+      {
+        singleLineText: {
+          value: "single line text",
+          type: "SINGLE_LINE_TEXT",
+        },
+        multiSelect: {
+          value: ["sample1", "sample2"],
+          type: "MULTI_SELECT",
+        },
+      },
+    ]);
+  });
+});
