+++ packages/data-loader/src/parsers/csvParser/__tests__/isImportSupportedFieldType.test.ts
@@ -1,0 +1,8 @@
+import { isImportSupportedFieldType } from "../isImportSupportedFieldType";
+
+describe("isImportSupportedFieldType", () => {
+  it("isImportSupportedFieldType", () => {
+    expect(isImportSupportedFieldType("SINGLE_LINE_TEXT")).toBe(true);
+    expect(isImportSupportedFieldType("RECORD_NUMBER")).toBe(false);
+  });
+});
