+++ packages/data-loader/src/parsers/csvParser/isImportSupportedFieldType.ts
@@ -1,0 +1,18 @@
+export const isImportSupportedFieldType = (fieldType: string) => {
+  return [
+    "SINGLE_LINE_TEXT",
+    "RADIO_BUTTON",
+    "MULTI_LINE_TEXT",
+    "NUMBER",
+    "RICH_TEXT",
+    "LINK",
+    "DROP_DOWN",
+    "CALC",
+    "UPDATED_TIME",
+    "CREATED_TIME",
+    "CREATOR",
+    "MODIFIER",
+    "MULTI_SELECT",
+    "CHECK_BOX",
+  ].includes(fieldType);
+};
